+++ tests/gif.js
@@ -6,18 +6,36 @@ group(() => {
 });
 
 group('getting time', () => {
-  bench('Date.now()', () => Date.now());
-  bench('Bun.nanoseconds()', () => Bun.nanoseconds());
-  bench('performance.now()', () => performance.now());
+  bench('Date.now()', () => {
+    Date.now();
+  });
+  bench('Bun.nanoseconds()', () => {
+    Bun.nanoseconds();
+  });
+  bench('performance.now()', () => {
+    performance.now();
+  });
 });
 
 group({ name: 'creating array', summary: false }, () => {
-  bench('new Array(0)', () => new Array(0));
-  bench('new Array(2 ** 2)', () => new Array(2 ** 2));
-  bench('new Array(2 ** 4)', () => new Array(2 ** 4));
-  bench('new Array(2 ** 8)', () => new Array(2 ** 8));
-  bench('new Array(2 ** 16)', () => new Array(2 ** 16));
-  bench('new Array(2 ** 20)', () => new Array(2 ** 20));
+  bench('new Array(0)', () => {
+    new Array(0);
+  });
+  bench('new Array(2 ** 2)', () => {
+    new Array(2 ** 2);
+  });
+  bench('new Array(2 ** 4)', () => {
+    new Array(2 ** 4);
+  });
+  bench('new Array(2 ** 8)', () => {
+    new Array(2 ** 8);
+  });
+  bench('new Array(2 ** 16)', () => {
+    new Array(2 ** 16);
+  });
+  bench('new Array(2 ** 20)', () => {
+    new Array(2 ** 20);
+  });
 });
 
 await run({ units: true });