+++ tests/test.ts
@@ -18,8 +18,12 @@ group(() => {
 
 group('group', () => {
   baseline('baseline', () => {});
-  bench('Date.now()', () => Date.now());
-  bench('performance.now()', () => performance.now());
+  bench('Date.now()', () => {
+    Date.now();
+  });
+  bench('performance.now()', () => {
+    performance.now();
+  });
 });
 
 group({ summary: false, after: () => {} }, () => {
@@ -28,8 +32,12 @@ group({ summary: false, after: () => {} }, () => {
 });
 
 group({ name: 'group2', summary: false }, () => {
-  bench('new Array(0)', () => new Array(0));
-  bench('new Array(1024)', () => new Array(1024));
+  bench('new Array(0)', () => {
+    new Array(0);
+  });
+  bench('new Array(1024)', () => {
+    new Array(1024);
+  });
 });
 
 bench('error', () => {